+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.26
 * Improvements in group call
 * Forward and View in chat action in Document Viewer
+* Improvements in check network
 
 ## 1.0.25
 * 🎉 Group call (beta)